# docs/postman/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/media/{filename}": {
      "delete": {
        "operationId": "MediaController_delete_v1",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
    "/api/v1/media/presigned/{filename}": {
      "post": {
        "operationId": "MediaController_getPresigned_v1",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
...
    "/api/v1/media/upload/presigned-url": {
      "post": {
        "operationId": "MediaController_createPresignedUrl_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUploadFileBodyDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
...
  },
  "info": {
    "title": "FastE API",
    "description": "The API for the ecommerce application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "payment-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    },
    "schemas": {
...
      "PresignedUploadFileBodyDTO": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string"
          },
          "filesize": {
            "type": "number",
            "maximum": 1048576,
            "exclusiveMaximum": false
          }
        },
        "required": [
          "filename",
          "filesize"
        ]
      },
...
    }
  }
}